Pull #9276: add one more closed to do issue pattern and fix violations
