+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/powermock/testmodules/CheckstyleAntTaskStub.java
@@ -1,0 +1,45 @@
+////////////////////////////////////////////////////////////////////////////////
+// checkstyle: Checks Java source code for adherence to a set of rules.
+// Copyright (C) 2001-2021 the original author or authors.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+////////////////////////////////////////////////////////////////////////////////
+
+package com.puppycrawl.tools.checkstyle.internal.powermock.testmodules;
+
+import static org.mockito.Mockito.when;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.powermock.api.mockito.PowerMockito;
+
+import com.puppycrawl.tools.checkstyle.ant.CheckstyleAntTask;
+
+public class CheckstyleAntTaskStub extends CheckstyleAntTask {
+
+    @Override
+    protected List<File> scanFileSets() {
+        final File mock = PowerMockito.mock(File.class);
+        // Assume that I/O error is happened when we try to invoke 'lastModified()' method.
+        final Exception expectedError = new RuntimeException("");
+        when(mock.lastModified()).thenThrow(expectedError);
+        final List<File> list = new ArrayList<>();
+        list.add(mock);
+        return list;
+    }
+
+}
